Refactor toss management and improve command handling
